Experiment with multi launch to enable keytool on windows
