+++ src/main/scala/io/iohk/ethereum/KeyTool.scala
@@ -1,0 +1,7 @@
+package io.iohk.ethereum
+
+
+object KeyTool {
+
+  def main(args: Array[String]): Unit = sun.security.tools.keytool.Main.main(args)
+}
